Move SpaceGroup interface to dedicated submodule
