+++ docs/api/autodoc/spglib.spg.rst
@@ -1,0 +1,8 @@
+spglib.spg module
+=================
+
+.. automodule:: spglib.spg
+   :members:
+   :private-members:
+   :show-inheritance:
+   :undoc-members:
